+++ lib/dimension.ts
@@ -1,38 +1,26 @@
 import { Block } from "./block.js";
 import { Entity } from "./entity.js";
+import { Wrapper } from "./wrapper.js";
 import { map } from "./iterable.js";
 import * as MC from "@minecraft/server";
 
-export class Dimension {
-    readonly #dimension: MC.Dimension;
-
-    /** The constructor is public only because of a language
-     * limitation. User code must never call it directly. */
-    public constructor(rawDimension: MC.Dimension) {
-        this.#dimension = rawDimension;
-    }
-
-    /// Package private
-    public get raw(): MC.Dimension {
-        return this.#dimension;
-    }
-
+export class Dimension extends Wrapper<MC.Dimension> {
     public get id(): string {
-        return this.#dimension.id;
+        return this.raw.id;
     }
 
     public getBlock(location: MC.Vector3): Block|undefined {
-        const raw = this.#dimension.getBlock(location);
+        const raw = this.raw.getBlock(location);
         return raw ? new Block(raw) : undefined;
     }
 
     public getEntities(opts?: MC.EntityQueryOptions): IterableIterator<Entity> {
-        return map(this.#dimension.getEntities(opts), raw => {
+        return map(this.raw.getEntities(opts), raw => {
             return new Entity(raw);
         });
     }
 
     public spawnEntity(identifier: string, location: MC.Vector3): Entity {
-        return new Entity(this.#dimension.spawnEntity(identifier, location));
+        return new Entity(this.raw.spawnEntity(identifier, location));
     }
 }