+++ lib/item/stack.ts
@@ -1,11 +1,9 @@
+import { Wrapper } from "../wrapper.js";
 import { ItemEnchants } from "./enchantment.js";
 import * as MC from "@minecraft/server";
 
-export class ItemStack {
-    readonly #itemStack: MC.ItemStack;
-
-    /** This overload is public only because of a language limitation. User
-     * code must never call it directly. */
+export class ItemStack extends Wrapper<MC.ItemStack> {
+    /** Package private: user code should not use this. */
     public constructor(rawItemStack: MC.ItemStack);
 
     /** Construct an item stack. */
@@ -13,38 +11,33 @@ export class ItemStack {
 
     public constructor(arg0: MC.ItemStack|MC.ItemType|string, ...rest: any[]) {
         if (arg0 instanceof MC.ItemStack) {
-            this.#itemStack = arg0;
+            super(arg0);
         }
         else {
-            this.#itemStack = new MC.ItemStack(arg0, ...rest);
+            super(new MC.ItemStack(arg0, ...rest));
         }
-    }
-
-    /** Package private: user code should not use this. */
-    get raw(): MC.ItemStack {
-        return this.#itemStack;
     }
 
     get amount(): number {
-        return this.#itemStack.amount;
+        return this.raw.amount;
     }
     set amount(n: number) {
-        this.#itemStack.amount = n;
+        this.raw.amount = n;
     }
 
     get typeId(): string {
-        return this.#itemStack.typeId;
+        return this.raw.typeId;
     }
 
     get lore(): string[] {
-        return this.#itemStack.getLore();
+        return this.raw.getLore();
     }
     set lore(lore: string[]) {
-        this.#itemStack.setLore(lore);
+        this.raw.setLore(lore);
     }
 
     get enchantments(): ItemEnchants {
         return new ItemEnchants(
-            this.#itemStack.getComponent("minecraft:enchantments") as MC.ItemEnchantsComponent);
+            this.raw.getComponent("minecraft:enchantments") as MC.ItemEnchantsComponent);
     }
 }